% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_predictions.R
\name{add_model_residuals}
\alias{add_model_residuals}
\title{Add model residuals}
\usage{
add_model_residuals(model = NULL, old_data = NULL)
}
\arguments{
\item{model}{A model object from `fit_model`}

\item{old_data}{The data set to which predicted values will be added.}
}
\value{
A data.frame object with residuals added.
}
\description{
A dplyr compatible convenience function to add residuals to a data set
}
\examples{
data("yields", package="manymodelr")
yields1 <- yields[1:50,]
yields2 <- yields[51:100,]
lm_model <- fit_model(yields1,"weight","height","lm") 
head(add_model_residuals(lm_model, yields2))
}
