% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_by_group.R
\name{agg_by_group}
\alias{agg_by_group}
\title{A convenient way to perform grouped operations}
\usage{
agg_by_group(data_set = NULL, my_formula = NULL, func = NULL, ...)
}
\arguments{
\item{data_set}{The data set for which correlations are required}

\item{my_formula}{A formula such as A~B where B is the grouping variable(normally a factor).
See examples below}

\item{func}{The kind of operation e.g sum,mean,min,max,manymodelr::get_mode}

\item{...}{Other arguments to 'aggregate' see ?aggregate for details}
}
\value{
A grouped data.frame object with results of the chosen operation.
}
\description{
This function performs operations by grouping the data.
}
\examples{
head(agg_by_group(airquality,.~Month,sum))
}
