% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drop_non_numeric}
\alias{drop_non_numeric}
\title{Drops non numeric columns from a data.frame object}
\usage{
drop_non_numeric(df)
}
\arguments{
\item{df}{A data.frame object for which non-numeric columns will be dropped}
}
\description{
Drops non numeric columns from a data.frame object
}
\examples{
 drop_non_numeric(data.frame(A=1:2, B=c("A", "B")))
}
