% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo.R
\name{expo}
\alias{expo}
\title{Convenience functions for use with get_exponent}
\usage{
expo(n)
}
\arguments{
\item{n}{The value to which a number should be raised. Useful for expo1}
}
\description{
Convenience functions for use with get_exponent
}
\details{
See ?get_exponent for details
}
\examples{
\dontrun{
square<-expo(2)
square(4)
expo1(2,3)}
}
