% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo.R
\name{expo1}
\alias{expo1}
\title{Convenience functions for use with get_exponent}
\usage{
expo1(y, x)
}
\arguments{
\item{y}{The value for which an exponential is required}

\item{x}{The power to which y should be raised}
}
\description{
Convenience functions for use with get_exponent
}
\details{
See ?get_exponent for detials
}
\examples{
\dontrun{
square<-expo(2)
square(4)
expo1(2,3)}
}
