% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_info.R
\name{extract_model_info}
\alias{extract_model_info}
\title{Extract important model attributes}
\usage{
extract_model_info(model_object, what)
}
\arguments{
\item{model_object}{A model object for example a linear model object, generalized linear model object,
analysis of variance object.}

\item{what}{character. The attribute you would like to obtain for instance p_value}
}
\value{
An object with the requested model info.
}
\description{
Provides a convenient way to extract any kind of model information from
common model objects
}
\examples{
#fit a linear model
lm_model <- lm(Sepal.Length ~ Petal.Length, iris)
extract_model_info(lm_model, "p_value")
}
