% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit and predict in a single function.}
\usage{
fit_model(df = NULL, yname = NULL, xname = NULL, modeltype = NULL, ...)
}
\arguments{
\item{df}{A data.frame object}

\item{yname}{The outcome variable}

\item{xname}{The predictor variable(s)}

\item{...}{Other arguments to specific model types.}

\item{model_type}{A character specifying the model type e.g lm for linear model}
}
\description{
Fit and predict in a single function.
}
\examples{
fit_model(iris,"Sepal.Length","Species","lm")
fit_model(iris,"Sepal.Width","Sepal.Length + Petal.Length + I(Petal.Width)**2","lm")
}
