% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_models}
\alias{fit_models}
\title{Fit several models with different response variables}
\usage{
fit_models(df = NULL, yname = NULL, xname = NULL, modeltype = NULL, ...)
}
\arguments{
\item{df}{A data.frame object}

\item{yname}{The outcome variable}

\item{xname}{The predictor variable(s)}

\item{modeltype}{A character specifying the model type e.g lm for linear model}

\item{...}{Other arguments to specific model types.}
}
\value{
A list of model objects that can be used later.
}
\description{
Fit several models with different response variables
}
\examples{
fit_models(df=iris,yname=c("Sepal.Length","Sepal.Width"),
xname="Petal.Length + Petal.Width",modeltype="lm")

}
