% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_data_Stats}
\alias{get_data_Stats}
\title{A pipe friendly way to get summary stats for exploratory data analysis}
\usage{
get_data_Stats(x, func)
}
\arguments{
\item{x}{The data for which stats are required}

\item{func}{The nature of function to apply}
}
\value{
A data.frame object showing the requested stats
}
\description{
A pipe friendly way to get summary stats for exploratory data analysis
}
\details{
A convenient wrapper especially useful for get_mode
}
\examples{
library(dplyr)
mtcars \%>\%
get_data_Stats(mean)
mtcars \%>\%
get_data_Stats(get_mode)
\dontrun{
get_data_Stats(airquality,min)
airquality\%>\%
get_data_Stats(get_mode)
}
}
