% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exponent.R
\name{get_exponent}
\alias{get_exponent}
\title{Get the exponent of any number or numbers}
\usage{
get_exponent(y, x)
}
\arguments{
\item{y}{The number for which an exponent is required}

\item{x}{The power to which y is raised}
}
\value{
A data.frame object showing the value,power and result
}
\description{
Get the exponent of any number or numbers
}
\details{
Depends on the expo and expo1 functions in expo
}
\examples{
df<-data.frame(A=c(1123,25657,3987))
get_exponent(df,3)
get_exponent(1:5, 6)
}
