% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_corr_.R
\name{get_var_corr_}
\alias{get_var_corr_}
\title{Get correlations for combinations}
\usage{
get_var_corr_(df, ...)
}
\arguments{
\item{df}{A `data.frame` object for which correlations are required in combinations.}

\item{...}{Other arguments to `get_var_corr`}
}
\value{
A data.frame object with combinations.
}
\description{
Get correlations for combinations
}
\details{
This function extends get_var_corr by providing an opportunity to get correlations
for combinations of variables. It is currently slow and may take up to a minute depending on system specifications.
}
\examples{
get_var_corr_(mtcars,exact=FALSE,method="spearman")
}
