% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeleR.R
\name{modeleR}
\alias{modeleR}
\title{Perform several kinds of models in one click}
\usage{
modeleR(df, yname, xname, modeltype, na.rm = F, ...)
}
\arguments{
\item{df}{The data for which analysis is required}

\item{yname}{The dependent variable}

\item{xname}{Theindependent variable. Supports formulae x1+x2+...}

\item{modeltype}{Currently one of lm and aov. Other models may work with inaccuracies}

\item{na.rm}{Logical. Should missing values be removed from analysis?}

\item{...}{Additional arguments to the modeltype}
}
\value{
A list containing summary stats and a data.frame object of some stats.
}
\description{
Perform several kinds of models in one click
}
\details{
This function provides a friendly way to perform any kind of model in one line.
}
\examples{
\dontrun{
modeleR(iris,Sepal.Length,Petal.Length,lm,na.rm=T)
}
}
