% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_model.R
\name{multi_model}
\alias{multi_model}
\title{Perform many models in a single command}
\usage{
multi_model(df, yname, xname, method, metric, control, ...)
}
\arguments{
\item{df}{The data holding the training dataset}

\item{yname}{The outcome variable}

\item{xname}{The predictor variable(s)}

\item{method}{One of the many methods available in caret}

\item{metric}{One of several metrics. Accuraciy,RMSE,MAE,etc}

\item{control}{See caret ?trainControl for details.}

\item{...}{Other arguments to caret's train function}
}
\value{
A data.frame object of summary statistics of the model types used
}
\description{
This function provides a convenient way to train several model types.
}
\details{
Most of the details of the parameters can be found in the caret package documentation.
This function is meant to help in exploratory analysis to make an informed choice of the best models
}
\examples{
\dontrun{
ctrl<-trainControl(method="cv",number=5)
multi_model(iris,"Species",".",c("rf","svmRadial","knn"),"Accuracy",ctrl)
my_grid<-expand.grid(.mtry=c(4:12))
multi_model(iris,"Species",".",c("rf"),"Accuracy",ctrl,tuneGrid=my_grid)
}
}
