% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace.R
\name{na_replace}
\alias{na_replace}
\title{Replace missing values}
\usage{
na_replace(df, how = NULL, value = NULL)
}
\arguments{
\item{df}{The data set(data.frame or vector) for which replacements are required}

\item{how}{How should missing values be replaced?  One of ffill, samples,value or any other known
method e.g mean, median, max ,min. The default is NULL meaning no  imputation is done. For character vectors,
the use of `get_mode` is also supported. No implementation for class factor(yet).}

\item{value}{If how is set to value, this allows the user to provide a specific fill value for the NAs.}
}
\value{
A data.frame object with missing values replaced.
}
\description{
Replace missing values
}
\details{
This function currently does not support grouping although this may be achieved with some
inaccuracies using grouping functions from other packages.
}
\examples{
na_replace(airquality,how="value", value="Missing")
}
