% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report_model}
\alias{report_model}
\title{Create a simplified report of a model's summary}
\usage{
report_model(model_object = NULL, response_name = "Score")
}
\arguments{
\item{model_object}{A model object}

\item{response_name}{Name of the response variable. Defaults to "Score".}
}
\value{
A data.frame object showing a simple model report that includes the
effect of each predictor variable on the response.
}
\description{
Create a simplified report of a model's summary
}
\examples{
models<-fit_models(df=yields,yname=c("height","yield"),xname="weight",
modeltype=c("lm", "glm"))
report_model(models[[2]][[1]])
}
