% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_percentile.R
\name{select_percentile}
\alias{select_percentile}
\title{Get the row corresponding to a given percentile}
\usage{
select_percentile(df, n)
}
\arguments{
\item{df}{The dataframe for which a percentile is required}

\item{n}{The percentile required eg 10th percentile}
}
\value{
A dataframe showing the row corresponding to a given percentile
}
\description{
Get the row corresponding to a given percentile
}
\examples{
select_percentile(iris,5)
\dontrun{
select_percentile(mtcars,1)
}
}
