% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model_predictions.R
\name{add_model_residuals}
\alias{add_model_residuals}
\title{Add model residuals}
\usage{
add_model_residuals(model, old_data)
}
\arguments{
\item{model}{A model object from `fit_model`}

\item{old_data}{The data set to which predicted values will be added.}
}
\value{
A data.frame object with residuals added.
}
\description{
A dplyr compatible convenience function to add residuals to a data set
}
\examples{
iris1 <- iris[1:50,]
iris2 <- iris[51:100,]
lm_model <- fit_model(iris1,"Sepal.Length","Sepal.Width","lm") 
head(add_model_residuals(lm_model,iris1))
}
