% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit and predict in a single function.}
\usage{
fit_model(df, yname, xname, modeltype, ...)
}
\arguments{
\item{df}{The data for which analysis is required}

\item{yname}{The dependent variable}

\item{xname}{The independent variable. Supports formulae x1+x2+...}

\item{modeltype}{Currently one of lm, glm and aov. Other models may work with inaccuracies}

\item{...}{Additional arguments to the modeltype}
}
\description{
Fit and predict in a single function.
}
\examples{
fit_model(iris,"Petal.Length","Sepal.Length","aov")
fit_model(iris,"Sepal.Width","Sepal.Length + Petal.Length + I(Petal.Width)**2","lm")
}
