% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeleR.R
\name{fit_model}
\alias{fit_model}
\title{Fit and predict in a single function.}
\usage{
fit_model(df, yname, xname, modeltype, na.rm = F, new_data, ...)
}
\arguments{
\item{df}{The data for which analysis is required}

\item{yname}{The dependent variable}

\item{xname}{The independent variable. Supports formulae x1+x2+...}

\item{modeltype}{Currently one of lm, glm and aov. Other models may work with inaccuracies}

\item{na.rm}{Logical. Should missing values be removed from analysis?}

\item{new_data}{A data.frame object for which new predictions are to be made}

\item{...}{Additional arguments to the modeltype}
}
\description{
Fit and predict in a single function.
}
\examples{
iris1<-iris[1:60,]
iris2<-iris[60:nrow(iris),]
fit_model(iris1,Sepal.Length,Petal.Length,
         lm,na.rm=TRUE,iris2) 
}
