% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_this.R
\name{get_this}
\alias{get_this}
\title{Helper function to easily access elements}
\usage{
get_this(what, where)
}
\arguments{
\item{what}{What do you want to extract from the `data.frame` or `list`? No quotes. See
examples below.}

\item{where}{Where do you want to get it from? Currently only supports `list`s
and `data.frame`objects.}
}
\description{
Helper function to easily access elements
}
\details{
This is a helper function useful if you would like to extract data from
the output of `multi_model_1`.
}
\examples{
my_list<-list(list(A=520),list(B=456))
get_this(A,my_list)
}
