% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_mean_na.R
\name{na_replace}
\alias{na_replace}
\title{Replace missing values}
\usage{
na_replace(df, how = NULL)
}
\arguments{
\item{df}{The data set for which replacements are required}

\item{how}{How should missing values be replaced?  One of ffill, samples or any other known
method e.g mean, median, max ,min. The default is NULL meaning no  imputation is done.}
}
\value{
A data.frame object with missing values replaced.
}
\description{
Replace missing values
}
\examples{
set.seed(520)
na_replace(airquality,how="samples")
}
