% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot a correlations matrix}
\usage{
plot_corr(data, x = "Comparison_Var", y = "Other_Var",
  xlabel = "Other_Variable", ylabel = "Comaprison_Variable",
  title = "Correlations Plot", plot_style = "circles",
  title_just = 0.5, round_values = TRUE, colour_by = "Correlation",
  decimals = 2, show_value = TRUE, size = 7, value_angle = 360,
  shape = 16, value_size = 3.5, value_col = "black", width = 7.5,
  custom_cols = c("white", "green2", "gray34"), ...)
}
\arguments{
\item{data}{The data to be plotted. A `data.frame` object produced
by `get_var_corr_`}

\item{x}{Value for the x axis. Defaults to "Comparison_Var"}

\item{y}{Values for the y axis. Defaults to "Other_Var."}

\item{xlabel}{label for the x axis}

\item{ylabel}{label for the y axis}

\item{title}{plot title.}

\item{plot_style}{One of squares and circles(currently).}

\item{title_just}{Justification of the title. Defaults to 0.5, title is centered.}

\item{round_values}{Logical. Should values be rounded off? Defaults to TRUE.}

\item{colour_by}{The column to use for coloring. Defaults to  "Correlation". Colour strength thus
indicates the strength of correlations.}

\item{decimals}{Numeric. To how many decimal places should the rounding be done?
Defaults to 2.}

\item{show_value}{Logical. Should correlation coefficients be shown on the plot?
Defaults to TRUE.}

\item{size}{Size of the circles for plot_style set to circles}

\item{value_angle}{What angle should the text be?}

\item{shape}{Values for the shape if plot_style is circles}

\item{value_size}{Size of the text.}

\item{value_col}{What colour should the text in the squares/circles be?}

\item{width}{width value for plot_style set to squares.}

\item{custom_cols}{A vector(length 2) of colors to use for the plot. The first colour specifies the lower end of the
correlations. The second specifies  the higher end.}

\item{...}{Other arguments to specific methods.}
}
\value{
A `ggplot2` object.
}
\description{
This function plots the results produced by
`get_var_corr_`.
}
\details{
This function uses `ggplot2` backend. `ggplot2` is thus required for the plots to work.
}
\examples{
# compute correlations
res<-get_var_corr_(mtcars)
# defaults
plot_corr(res)
plot_corr(res, decimals = 2, plot_style = "circles",
show_value = TRUE, size = 7,
value_angle =360, colour_by = "Correlation",
shape = 16,
value_size = 3.5,
width = 7.5,
custom_cols = c("white","green2",
"gray34"),
             value_col = "black",
             x="Other_Var",y="Comparison_Var",
             xlabel = "Other_Variable",
             ylabel="Comaprison_Variable")

}
