% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_percentile.R
\name{select_percentile}
\alias{select_percentile}
\title{Get the row corresponding to a given percentile}
\usage{
select_percentile(df, percentile, get_all = TRUE, descend = FALSE)
}
\arguments{
\item{df}{A `data.frame` object for which a percentile is required. Other data structures are not yet supported.}

\item{percentile}{The percentile required eg 10 percentile}

\item{get_all}{Logical. Should all variables be used? Currently only supports
using all variables.}

\item{descend}{Logical. Should the data be arranged in descending order? Defaults
to FALSE.}
}
\value{
A dataframe showing the row corresponding to the required percentile.
}
\description{
Get the row corresponding to a given percentile
}
\details{
Returns the value corresponding to a percentile. 
Returns mean values if the position of the percentile is  whole number.
Values are sorted in ascending order. You can change this by setting descend to TRUE.
}
\examples{
select_percentile(iris,5)
}
