% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_percentile.R
\name{select_percentile}
\alias{select_percentile}
\title{Get the row corresponding to a given percentile}
\usage{
select_percentile(df, percentile, get_all = T, ascend = FALSE)
}
\arguments{
\item{df}{The dataframe for which a percentile is required}

\item{percentile}{The percentile required eg 10 percentile}

\item{get_all}{Logical. Should all variables be used? Currenty only supports
using all variables.}

\item{ascend}{Logical. Should the data be arranged in ascending order? Defaults
to FALSE.}
}
\value{
A dataframe showing the row corresponding to the required percentile.
 @examples
select_percentile(iris,5,)
}
\description{
Get the row corresponding to a given percentile
}
